+++ cmd/data-node/commands/networkhistory/fetch.go
@@ -64,7 +64,7 @@ func (cmd *fetchCmd) Execute(args []string) error {
 		return fmt.Errorf("failed to fix config:%w", err)
 	}
 
-	err = verifyChainID(cmd.SQLStore.ConnectionConfig, cmd.ChainID)
+	err = verifyChainID(ctx, cmd.SQLStore.ConnectionConfig, cmd.ChainID)
 	if err != nil {
 		return fmt.Errorf("failed to verify chain id:%w", err)
 	}
@@ -99,8 +99,8 @@ func (cmd *fetchCmd) Execute(args []string) error {
 	return nil
 }
 
-func verifyChainID(connConfig sqlstore.ConnectionConfig, chainID string) error {
-	connSource, err := sqlstore.NewTransactionalConnectionSource(logging.NewTestLogger(), connConfig)
+func verifyChainID(ctx context.Context, connConfig sqlstore.ConnectionConfig, chainID string) error {
+	connSource, err := sqlstore.NewTransactionalConnectionSource(ctx, logging.NewTestLogger(), connConfig)
 	if err != nil {
 		return fmt.Errorf("failed to create new transactional connection source: %w", err)
 	}