+++ datanode/sqlstore/fees_stats.go
@@ -56,12 +56,7 @@ func NewFeesStats(src *ConnectionSource) *FeesStats {
 func (rfs *FeesStats) AddFeesStats(ctx context.Context, stats *entities.FeesStats) error {
 	defer metrics.StartSQLQuery("FeesStats", "AddFeesStats")()
 
-	ctx, err := rfs.WithTransaction(ctx)
-	if err != nil {
-		return fmt.Errorf("could not initialise transaction: %w", err)
-	}
-
-	if _, err := rfs.Connection.Exec(
+	if _, err := rfs.Exec(
 		ctx,
 		`INSERT INTO fees_stats(
 			   market_id,
@@ -86,24 +81,16 @@ func (rfs *FeesStats) AddFeesStats(ctx context.Context, stats *entities.FeesStat
 		stats.MakerFeesGenerated,
 		stats.VegaTime,
 	); err != nil {
-		_ = rfs.Rollback(ctx)
 		return fmt.Errorf("could not execute insertion in `fees_stats`: %w", err)
 	}
 
-	partiesStats := computePartiesStats(stats)
-
 	batcher := NewListBatcher[*feesStatsForPartyRow]("fees_stats_by_party", feesStatsByPartyColumn)
+	partiesStats := computePartiesStats(stats)
 	for _, s := range partiesStats {
 		batcher.Add(s)
 	}
-
-	if _, err := batcher.Flush(ctx, rfs.Connection); err != nil {
-		_ = rfs.Rollback(ctx)
-		return fmt.Errorf("could not flush the batch insertion in `fees_stats_by_party`: %w", err)
-	}
-
-	if err := rfs.Commit(ctx); err != nil {
-		return fmt.Errorf("an error occurred during transactions commit: %w", err)
+	if _, err := batcher.Flush(ctx, rfs.ConnectionSource); err != nil {
+		return err
 	}
 
 	return nil
@@ -143,7 +130,7 @@ func (rfs *FeesStats) StatsForParty(ctx context.Context, partyID entities.PartyI
 	)
 
 	var rows []feesStatsForPartyRow
-	if err := pgxscan.Select(ctx, rfs.Connection, &rows, query, args...); err != nil {
+	if err := pgxscan.Select(ctx, rfs.ConnectionSource, &rows, query, args...); err != nil {
 		return nil, err
 	}
 
@@ -202,7 +189,7 @@ func (rfs *FeesStats) GetFeesStats(ctx context.Context, marketID *entities.Marke
 
 	query = fmt.Sprintf("%s order by market_id, asset_id, epoch_seq desc", query)
 
-	if err = pgxscan.Select(ctx, rfs.Connection, &stats, query, args...); err != nil {
+	if err = pgxscan.Select(ctx, rfs.ConnectionSource, &stats, query, args...); err != nil {
 		return nil, err
 	}