+++ datanode/sqlstore/referral_programs_test.go
@@ -121,7 +121,7 @@ func TestReferralPrograms_AddReferralProgram(t *testing.T) {
 		require.NoError(t, err)
 
 		var got []entities.ReferralProgram
-		err = pgxscan.Select(ctx, connectionSource.Connection, &got, "SELECT * FROM referral_programs")
+		err = pgxscan.Select(ctx, connectionSource, &got, "SELECT * FROM referral_programs")
 		require.NoError(t, err)
 		require.Len(t, got, 1)
 		assert.Equal(t, *want, got[0])
@@ -130,7 +130,7 @@ func TestReferralPrograms_AddReferralProgram(t *testing.T) {
 		err = rs.AddReferralProgram(ctx, want2)
 		require.NoError(t, err)
 
-		err = pgxscan.Select(ctx, connectionSource.Connection, &got, "SELECT * FROM referral_programs")
+		err = pgxscan.Select(ctx, connectionSource, &got, "SELECT * FROM referral_programs")
 		require.NoError(t, err)
 		require.Len(t, got, 2)
 		wantAll := []entities.ReferralProgram{*want, *want2}
@@ -238,7 +238,7 @@ func TestReferralPrograms_UpdateReferralProgram(t *testing.T) {
 		require.NoError(t, err)
 
 		var got []entities.ReferralProgram
-		err = pgxscan.Select(ctx, connectionSource.Connection, &got, "SELECT * FROM referral_programs")
+		err = pgxscan.Select(ctx, connectionSource, &got, "SELECT * FROM referral_programs")
 		require.NoError(t, err)
 
 		require.Len(t, got, 1)
@@ -249,7 +249,7 @@ func TestReferralPrograms_UpdateReferralProgram(t *testing.T) {
 		err = rs.UpdateReferralProgram(ctx, wantUpdated)
 		require.NoError(t, err)
 
-		err = pgxscan.Select(ctx, connectionSource.Connection, &got, "SELECT * FROM referral_programs")
+		err = pgxscan.Select(ctx, connectionSource, &got, "SELECT * FROM referral_programs")
 		require.NoError(t, err)
 
 		require.Len(t, got, 2)
@@ -260,7 +260,7 @@ func TestReferralPrograms_UpdateReferralProgram(t *testing.T) {
 
 	t.Run("The current_referral view should list the updated referral program record", func(t *testing.T) {
 		var got []entities.ReferralProgram
-		err := pgxscan.Select(ctx, connectionSource.Connection, &got, "SELECT * FROM current_referral_program")
+		err := pgxscan.Select(ctx, connectionSource, &got, "SELECT * FROM current_referral_program")
 		require.NoError(t, err)
 		require.Len(t, got, 1)
 		assert.Equal(t, *wantUpdated, got[0])
@@ -294,13 +294,13 @@ func TestReferralPrograms_EndReferralProgram(t *testing.T) {
 		ended.EndedAt = &endTime
 
 		var got []entities.ReferralProgram
-		err = pgxscan.Select(ctx, connectionSource.Connection, &got, "SELECT * FROM referral_programs order by vega_time")
+		err = pgxscan.Select(ctx, connectionSource, &got, "SELECT * FROM referral_programs order by vega_time")
 		require.NoError(t, err)
 		require.Len(t, got, 3)
 		wantAll := []entities.ReferralProgram{*started, *updated, *ended}
 		assert.Equal(t, wantAll, got)
 
-		err = pgxscan.Select(ctx, connectionSource.Connection, &got, "SELECT * FROM current_referral_program")
+		err = pgxscan.Select(ctx, connectionSource, &got, "SELECT * FROM current_referral_program")
 		require.NoError(t, err)
 		require.Len(t, got, 1)
 		assert.Equal(t, *ended, got[0])