+++ datanode/sqlstore/rewards.go
@@ -55,7 +55,7 @@ func NewRewards(ctx context.Context, connectionSource *ConnectionSource) *Reward
 func (rs *Rewards) fetchRunningTotals(ctx context.Context) {
 	query := `SELECT * FROM current_game_reward_totals`
 	var totals []entities.RewardTotals
-	err := pgxscan.Select(ctx, rs.Connection, &totals, query)
+	err := pgxscan.Select(ctx, rs.ConnectionSource, &totals, query)
 	if err != nil && !pgxscan.NotFound(err) {
 		panic(fmt.Errorf("could not retrieve game reward totals: %w", err))
 	}
@@ -73,7 +73,7 @@ func (rs *Rewards) fetchRunningTotals(ctx context.Context) {
 
 func (rs *Rewards) Add(ctx context.Context, r entities.Reward) error {
 	defer metrics.StartSQLQuery("Rewards", "Add")()
-	_, err := rs.Connection.Exec(ctx,
+	_, err := rs.Exec(ctx,
 		`INSERT INTO rewards(
 			party_id,
 			asset_id,
@@ -109,7 +109,7 @@ func (rs *Rewards) Add(ctx context.Context, r entities.Reward) error {
 		rs.runningTotalsQuantum[gID][r.PartyID] = rs.runningTotalsQuantum[gID][r.PartyID].Add(r.QuantumAmount)
 
 		defer metrics.StartSQLQuery("GameRewardTotals", "Add")()
-		_, err = rs.Connection.Exec(ctx, `INSERT INTO game_reward_totals(
+		_, err = rs.Exec(ctx, `INSERT INTO game_reward_totals(
 			game_id,
 			party_id,
 			asset_id,
@@ -155,7 +155,7 @@ type scannedRewards struct {
 func (rs *Rewards) GetAll(ctx context.Context) ([]entities.Reward, error) {
 	defer metrics.StartSQLQuery("Rewards", "GetAll")()
 	scanned := []scannedRewards{}
-	err := pgxscan.Select(ctx, rs.Connection, &scanned, `SELECT * FROM rewards;`)
+	err := pgxscan.Select(ctx, rs.ConnectionSource, &scanned, `SELECT * FROM rewards;`)
 	if err != nil {
 		return nil, err
 	}
@@ -166,7 +166,7 @@ func (rs *Rewards) GetByTxHash(ctx context.Context, txHash entities.TxHash) ([]e
 	defer metrics.StartSQLQuery("Rewards", "GetByTxHash")()
 
 	scanned := []scannedRewards{}
-	err := pgxscan.Select(ctx, rs.Connection, &scanned, `SELECT * FROM rewards WHERE tx_hash = $1`, txHash)
+	err := pgxscan.Select(ctx, rs.ConnectionSource, &scanned, `SELECT * FROM rewards WHERE tx_hash = $1`, txHash)
 	if err != nil {
 		return nil, err
 	}
@@ -199,7 +199,7 @@ func (rs *Rewards) GetByCursor(ctx context.Context,
 	}
 
 	scanned := []scannedRewards{}
-	if err := pgxscan.Select(ctx, rs.Connection, &scanned, query, args...); err != nil {
+	if err := pgxscan.Select(ctx, rs.ConnectionSource, &scanned, query, args...); err != nil {
 		return nil, entities.PageInfo{}, fmt.Errorf("querying rewards: %w", err)
 	}
 
@@ -218,7 +218,7 @@ func (rs *Rewards) GetSummaries(ctx context.Context,
 
 	summaries := []entities.RewardSummary{}
 	defer metrics.StartSQLQuery("Rewards", "GetSummaries")()
-	err := pgxscan.Select(ctx, rs.Connection, &summaries, query, args...)
+	err := pgxscan.Select(ctx, rs.ConnectionSource, &summaries, query, args...)
 	if err != nil {
 		return nil, fmt.Errorf("querying rewards: %w", err)
 	}
@@ -247,7 +247,7 @@ func (rs *Rewards) GetEpochSummaries(ctx context.Context,
 	var summaries []entities.EpochRewardSummary
 	defer metrics.StartSQLQuery("Rewards", "GetEpochSummaries")()
 
-	if err = pgxscan.Select(ctx, rs.Connection, &summaries, query, args...); err != nil {
+	if err = pgxscan.Select(ctx, rs.ConnectionSource, &summaries, query, args...); err != nil {
 		return nil, pageInfo, fmt.Errorf("querying epoch reward summaries: %w", err)
 	}